Get tests working w/ grunt-esri-slurp to download a local copy of JSAPI
I can't get the Intern to work w/ a CDN hosted ver of the JSAPI.
The browser test runner is very flakey (throws 404 error on host-node?dojo/node.js) half the time. So, I used slurp to download a local copy of of the JSAPI an pointed the Intern at the local copy. Now the browser tests pass every time.

This required a few significant changes:

- updated to JSAPI 3.10
- added grunt-esri-slurp as dev dependency
- downloading a local copy of Dojo w/ bower
- downloading a local copy of JSAPI w/ slurp
- updated to latest version of Intern (2.0.3)
- updated README instructions

If you like this idea and merge this PR, then I can take on #6.